Add context relevancy metric for RAG answers
This ports over the functionality from the prototype which can be found here:
https://github.com/alphagov/govuk-chat-monitoring-prototyping/tree/fdba73eb2469f81b6266dc9a3b1ab1aabda63d85/src/evaluation/contextquality

I've refactored the functionality into a new metric within the custom_deepeval
metrics for RAG answers. It's largely based off the conventions in the
factual correctness metric.

The metric takes the retrived contexts and a user question, breaks the
question down into needs, breaks the contexts down into facts and then
evaluates how well the facts in the contexts meet the needs in the question.
It returns a score between 0 and 1 indicating the relevancy of the contexts
to the question.

There's quite a lot going on in this module. I've tried to keep the tests
reasonably comprehensive to cover the various parts of the functionality,
but have tried to only test the public interfaces where possible.