Extract functionality out into schema methods
There's a good amount of logic that can be extracted and unit tested
in the schema classes.

This moves stuff like aggregating attributes from lists of objects into
methods on the schema classes themselves. As well as things like calculating
scores based on the verdicts.

Nick mentioned in some PR feedback that the extracted truths should contain
the context as well as the facts. This is now represented as a list of dicts
in the schema, rather than a list of strings. It has the following shape:

```
[
    {"context": "ctx1", "truths": ["fact1", "fact2"]},
    {"context": "ctx2", "truths": ["fact3"]},
]

```