Implement coherence metric
This commit ports the coherence metric from the prototyping app: https://github.com/alphagov/govuk-chat-monitoring-prototyping/tree/8a740e3de2ea5abcc052ed5ce4daeb8a039fd71b/src/evaluation/coherence

Coherence was initially implemented in the prototype using the DeepEval
G-Eval metric. However, following discussion with @kemi-bower and
investigation of the G-Eval implementation, it was decided that G-Eval
didn't offer us much in this case, and that it would be straightforward
to implement it as a custom metric (in the same vein as the factual
correctness and context relevancy metrics). This allows us to stay
closer to established patterns, and avoid adding new abstractions to the
codebase simply for the sake of this single metric.

I've stayed somewhat close to the G-Eval implementation of the metric
but have refined it to what is needed in our context.

Co-authored-by: Kemi Bower <kemi.bower@digital.cabinet-office.gov.uk>